added a basic camera
The camera isn't really a camera, it's just a class that abstracts away the fact that "moving the camera right" is equivalent to moving the entire universe left. It also makes the camera move smoothly, instead of putting it in a new place instantly (which has NOT BEEN TESTED yet btw).